Update LICENSE
Addition of Missing Topics

In this commit, we have updated the "ETHICAL-FREEDOM LICENSE" to include the missing topics that were not explicitly covered in the original license. These additions enhance the comprehensiveness of the license by addressing key aspects such as commercial use, warranty and liability, termination, attribution, jurisdiction and governing law, version updates, responsibility for compliance, compatibility with other licenses, and international considerations. The modifications ensure that the license provides a more comprehensive framework for the use of the software and its derivatives while emphasizing ethical considerations and legal compliance.